Crate srt2webvtt

Structs

Cue
One cue.

SrtParser

WebVTTParser
A parser of a WebVTT stream.

Enums

Delta
A delta duration to apply on a cue's time code.

Format
The crate supported formats for input or output stream.

Functions

convert
Convert cues from the input, apply delta duration and save it.

convert_output
Apply the delta time to all input cues and save them into the output_writer.

srt_out
Write all Cues from the input Iterator into the write W. Use SRT subtitle format. Return the number fo writed cue.

webvtt_out
Write all Cues from the input Iterator into the write W. Use WebVTT subtitle format. Return the number fo writed cue.